modified code structure to follow Google C++ Style Guide
