+++ src/pub_main.cpp
@@ -1,0 +1,15 @@
+
+
+
+
+#include "../include/beginner_tutorials/publisher_member_function.hpp"
+
+
+int main(int argc, char * argv[])
+{
+  rclcpp::init(argc, argv);
+  rclcpp::spin(std::make_shared<MinimalPublisher>());
+  rclcpp::shutdown();
+  return 0;
+}
+
